feat: add function to remove trailing spaces chore: move Utils.js to appropriate lovation and update imports
